variable lenght arguments for pytorch network
- now both fit and predict takes a variable argument lenght
- now the tensor type conversion is perform by the pytorch network, not
by the interface

+++ examples/atari_dqn.py
@@ -57,7 +57,7 @@ class Network(nn.Module):
         if action is None:
             return q
         else:
-            q_acted = torch.squeeze(q.gather(1, action))
+            q_acted = torch.squeeze(q.gather(1, action.long()))
 
             return q_acted
